Ported acerhk to kernel version 3.16. Probably works with kernel versions > 3.10.
Things to consider: this is a dirty hack to make it work for me. I tried to keep backwards compatibility, but neither tested it nor put too much thought into it. Also there is a workaround regarding ebp not being allowed in asm, I did not test it and am not knowing what I was doing there, I just copied it from some explanation dealing with a error message I got when compiling.

Therefore although it works for me.

THIS CODE IS PROBABLY FAULTY AND MAY BREAK YOUR SYSTEM OR POSE SEVERE SECURITY ISSUES. USE AT YOUR OWN RISK.

	modified:       Makefile
	modified:       acerhk.c